requirement.txt
numpy
faiss-cpu
torch
transformers
sentence-transformers
spacy
pyspellchecker